# docs/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Asmblyr Core HTTP API",
    "version": "0.0.0",
    "description": "Все статические маршруты Core и их требования доступа. Часть операций имеет только route-only описание; спецификация не является полной схемой для генерации клиента. OAuth protocol и динамические маршруты плагинов описаны отдельно."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Локальный Core"
    }
  ],
  "paths": {
...
    "/auth/sso/{provider}/callback": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "post__auth_sso_provider_callback",
        "summary": "POST /auth/sso/:provider/callback",
        "description": "Проверка SSO flow, state/PKCE и ответа провайдера. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/auth/sso/routes.ts",
        "security": [],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/auth/sso/routes.ts"
      }
    },
    "/auth/sso/{provider}/start": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "post__auth_sso_provider_start",
        "summary": "POST /auth/sso/:provider/start",
        "description": "Начало SSO; привязка требует действующей пользовательской сессии. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/auth/sso/routes.ts",
        "security": [],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/auth/sso/routes.ts"
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "get__users",
        "summary": "GET /users",
        "description": "Человек: users/read или update ИЛИ policies/read или update; superuser bypass. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/auth/user-routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/auth/user-routes.ts"
      },
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "post__users",
        "summary": "POST /users",
        "description": "Человек: users/update; superuser bypass. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/auth/user-routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/auth/user-routes.ts"
      }
    },
    "/users/{id}/access": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "get__users_id_access",
        "summary": "GET /users/:id/access",
        "description": "Человек: users/read ИЛИ update; superuser bypass. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/auth/user-routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/auth/user-routes.ts"
      }
    },
    "/users/{id}/delegation": {
      "put": {
        "tags": [
          "users"
        ],
        "operationId": "put__users_id_delegation",
        "summary": "PUT /users/:id/delegation",
        "description": "Только человек-superuser: явный набор готовых политик, которые пользователь вправе назначать. Источник: apps/core/src/auth/user-routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "policyIds": {
                          "type": "array",
                          "maxItems": 1000,
                          "uniqueItems": true,
                          "items": {
                            "type": "string",
                            "format": "uuid"
                          }
                        }
                      },
                      "required": [
                        "policyIds"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Ошибка проверки, авторизации или исполнения; код HTTP определяет результат."
          }
        },
        "x-contract-level": "documented",
        "x-source": "apps/core/src/auth/user-routes.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "policyIds": {
                    "type": "array",
                    "maxItems": 1000,
                    "uniqueItems": true,
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                },
                "required": [
                  "policyIds"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/users/{id}/invitation": {
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "post__users_id_invitation",
        "summary": "POST /users/:id/invitation",
        "description": "Человек: superuser ИЛИ users/update; повторное приглашение только в пределах разрешённого набора, без доступа к аккаунтам с собственным делегированием. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/auth/user-routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/auth/user-routes.ts"
      }
    },
    "/users/{id}/recovery": {
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "post__users_id_recovery",
        "summary": "POST /users/:id/recovery",
        "description": "Свежий человек-superuser; только другой активный обычный пользователь. Одноразовая ссылка на 30 минут, аудит; не делегируется. Источник: apps/core/src/auth/user-routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "invitationToken": {
                          "type": "string"
                        },
                        "expiresAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "invitationToken",
                        "expiresAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "x-contract-level": "documented",
        "x-source": "apps/core/src/auth/user-routes.ts"
      }
    },
...
    "/users/me/identities": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "get__users_me_identities",
        "summary": "GET /users/me/identities",
        "description": "Человек, только собственный профиль/ресурс; workspace дополнительно проверяет видимость. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/auth/sso/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/auth/sso/routes.ts"
      }
    },
    "/users/me/identities/{id}": {
      "delete": {
        "tags": [
          "users"
        ],
        "operationId": "delete__users_me_identities_id",
        "summary": "DELETE /users/me/identities/:id",
        "description": "Человек, только собственный профиль/ресурс; workspace дополнительно проверяет видимость. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/auth/sso/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/auth/sso/routes.ts"
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Непрозрачный access token пользователя или сервиса. Разделы настроек доступны только человеку."
      }
    }
  },
...
}